+++ infra/controller.js
@@ -1,0 +1,26 @@
+import { InternalServerError, MethodNotAllowedError } from "infra/errors";
+
+function onNoMatchHandler(request, response) {
+	const publicErrorObject = new MethodNotAllowedError();
+	response.status(publicErrorObject.statusCode).json(publicErrorObject);
+}
+
+function onErrorHandler(error, request, response) {
+	const publicErrorObject = new InternalServerError({
+		statusCode: error.statusCode,
+		cause: error,
+	});
+
+	console.error(publicErrorObject);
+
+	response.status(publicErrorObject.statusCode).json(publicErrorObject);
+}
+
+const controller = {
+	errorHandlers: {
+		onNoMatch: onNoMatchHandler,
+		onError: onErrorHandler,
+	},
+};
+
+export default controller;
