+++ pages/api/v1/status/index.js
@@ -1,26 +1,12 @@
 import { createRouter } from "next-connect";
 import database from "infra/database.js";
-import { InternalServerError, MethodNotAllowedError } from "infra/errors";
+import controller from "infra/controller.js";
 
 const router = createRouter();
 
 router.get(getHandler);
 
-export default router.handler({ onNoMatch: onNoMatchHandler, onError: onErrorHandler });
-
-function onNoMatchHandler(request, response) {
-	const publicErrorObject = new MethodNotAllowedError();
-	response.status(publicErrorObject.statusCode).json(publicErrorObject);
-}
-
-function onErrorHandler(error, request, response) {
-	const publicErrorObject = new InternalServerError({ cause: error });
-
-	console.log("\n Erro dentro do catch do next-connect:");
-	console.error(publicErrorObject);
-
-	response.status(500).json(publicErrorObject);
-}
+export default router.handler(controller.errorHandlers);
 
 async function getHandler(request, response) {
 	const updatedAt = new Date().toISOString();
